ocaml for low-level system functionality
To my great surprise it is easier to implement low-level Unix helpers
in Ocaml than in Go. Lwt library in Ocaml with its promises that can be
canceled is really nice for this stuff. In Go one cannot cancel threads
which leads to unexpectedly complex code to ensure proper shutdown
using multiple channels.